+++ src/spatRaster.h
@@ -595,9 +595,9 @@ class SpatRaster {
 		std::vector<double> bilinearCells(const std::vector<double> &x, const std::vector<double> &y);
 		std::vector<double> fourCellsFromXY(const std::vector<double> &x, const std::vector<double> &y);
 
-
 		SpatRaster buffer(double d, double background, SpatOptions &opt);
 		SpatRaster clamp(std::vector<double> low, std::vector<double> high, bool usevalue, SpatOptions &opt);
+		SpatRaster clamp_raster(SpatRaster &x, SpatRaster &y, std::vector<double> low, std::vector<double> high, bool usevalue, SpatOptions &opt);
 		SpatRaster clamp_ts(bool min, bool max, SpatOptions &opt);
 
 		SpatRaster combineCats(SpatRaster x, SpatOptions &opt);